small fixes

+++ howdy/views.py
@@ -194,7 +194,7 @@ def SalaryPrediction(request, *args, **kwargs):
         # predicted = int(queryset) * 69
 
         to_predict = []
-        to_predict.append(int(queryset))
+        to_predict.append(float(queryset))
         to_predict = np.array(to_predict)
 
         predicted = poly_regressor.predict(regressor.fit_transform(to_predict.reshape(-1,1)))
